athena_app_cmdb/requirements/base.txt
django>=3.2.2
DjangoRestFramework
psycopg2
gunicorn
gevent
requests
redis
django-redis
django-session-timeout
pyyaml
pydash
requests-oauthlib
yaql
openpyxl
django-widget-tweaks
django-compressor
jsonpatch
djangorestframework_simplejwt
tblib
django-widget-tweaks
django-bootstrap4
django-bootstrap-modal-forms
json-log-formatter
django-auth-ldap
python-ldap>=3.0
ldap3
django-json-widget
django-currentuser
django-admin-autocomplete-filter